Update widths to maxwidth to be more responsive
I couldn't figure out how the source was setting the width of the elements so I took a look at their code. They are using a maxwidth on each container with a padding on either side. So when the screen is made narrow, the items will keep their size until a certain point, then scale down. I copied this technique on my page so now the widths change the same way they do in the source. The next step is to set up the breakpoints and their changes. Almost done!